♻️ (keyring): Split authentication device actions
