+++ packages/trusted-apps/ledger-keyring-protocol/src/internal/app-binder/task/ExtractEncryptionKeyTask.ts
@@ -1,0 +1,26 @@
+import { LKRPUnknownError } from "@api/app-binder/Errors";
+import { type Keypair } from "@api/index";
+import { type LKRPBlockStream } from "@internal/utils/LKRPBlockStream";
+
+export type ExtractEncryptionKeyTaskInput = {
+  applicationStream: LKRPBlockStream;
+  keypair: Keypair;
+};
+
+export class ExtractEncryptionKeyTask {
+  async run(keypair: Keypair, stream: LKRPBlockStream) {
+    // TODO additional derivations should be supported:
+    // https://github.com/LedgerHQ/ledger-live/blob/develop/libs/hw-ledger-key-ring-protocol/src/Device.ts#L216...L226
+    // Probably not needed for Ledger Sync
+    return Promise.resolve(
+      stream
+        .getPublishedKey(keypair)
+        .map((key) => key.privateKey)
+        .toEither(
+          new LKRPUnknownError(
+            "There is no encryption key for the current member in the application stream.",
+          ),
+        ),
+    );
+  }
+}
